+++ internal/cli/cli.go
@@ -66,15 +66,6 @@ func Main(version string, buildDate string) int {
 		return 1
 	}
 
-	if rootFlags.version {
-		if err := cmd.PrintVersionMessage(rootCmd); err != nil {
-			log.Errorf("print version error: %s", err)
-			return 1
-		}
-
-		return 0
-	}
-
 	debugLevel := env.SetDebugLevel(rootFlags.debug)
 
 	if debugLevel > 0 {
@@ -106,17 +97,6 @@ func Main(version string, buildDate string) int {
 
 		if err != nil {
 			log.Errorf("can not create lets.yaml: %s", err)
-			return 1
-		}
-
-		return 0
-	}
-
-	showUsage := rootFlags.help || (command.Name() == "help" && len(args) == 0) || (len(os.Args) == 1)
-
-	if showUsage {
-		if err := cmd.PrintRootHelpMessage(rootCmd); err != nil {
-			log.Errorf("print help error: %s", err)
 			return 1
 		}
 
@@ -178,9 +158,12 @@ func getExitCode(err error, defaultCode int) int {
 	return defaultCode
 }
 
-// do not fail on config error if it is help (-h, --help), --init, completion, or lets self.
+// do not fail on config error if it is help (-h, --help), --version, --init, completion, or lets self.
 func failOnConfigError(root *cobra.Command, current *cobra.Command, rootFlags *flags) bool {
-	return (root.Flags().NFlag() == 0 && !allowsMissingConfig(current)) && !rootFlags.help && !rootFlags.init
+	return (root.Flags().NFlag() == 0 && !allowsMissingConfig(current)) &&
+		!rootFlags.help &&
+		!rootFlags.version &&
+		!rootFlags.init
 }
 
 func allowsMissingConfig(current *cobra.Command) bool {